[Refactor] - Move AudioDetails files to ':media_ui' module
