+++ media_ui/build.gradle
@@ -5,6 +5,9 @@ dependencies {
     kapt libs.dagger_compiler
     kapt libs.dagger_android_compiler
 
+    implementation libs.exo_player_core
+    implementation libs.exo_player_ui
+
     implementation libs.constraint_layout
     implementation project(':viewmodel')
     implementation libs.rx_helpers
@@ -15,4 +18,5 @@ dependencies {
     implementation project(':navigation')
     implementation project(':viewmodel')
     implementation project(':adapter')
+    implementation project(':image')
 }
