+++ media_ui/src/main/res/layout/exo_playback_control_view.xml
