+++ media_ui/src/main/res/layout/fragment_audio_details.xml
